mctp: fix payload size check corner case
current payload size need to be at least 5 bytes to cover random seed
for receiver to check corresponding random data.

Previous check use modulas to check payload size, which may pad payload
size incorrectly if the size is multiple of mtu size.

add check so that payload size is at least 5 bytes, and not padded if it
is a multiple of mtu size.

Signed-off-by: aspeedyh <yh_chung@aspeedtech.com>
Change-Id: Ifbf9fffb7d66c03155d9852be7d38eaedbb13786